Archive old JS; start shiny, new JS
**Why**:
- we had a lot of older and no-longer-necessary JS
- we should be taking advantage of [ES6](http://es6-features.org)

**How**:
- add `browserify-rails` gem to transpile ES6 for older browser
compatibility
- clean up / remove older JS files, and start to add back the pieces
we need
