+++ .eslintrc
@@ -1,5 +1,4 @@
 {
-  // All rules should be disabled or they should produce errors. No warnings.
   'extends': 'airbnb',
   'parser': 'babel-eslint',
   'env': {
@@ -8,9 +7,6 @@
     'es6': true,
     'mocha': true
   },
-  'plugins': [
-    'react'
-  ],
   'globals': {
     '$': true,
     '__DEV__': true,
@@ -20,45 +16,7 @@
     'expect': true,
     'fixture': true
   },
-  'rules': {
-    // Override airbnb style.
-    'arrow-body-style': 0,  // Leave bracing to code reviewer discretion.
-    'camelcase': [2, {'properties': 'always'}],
-    'comma-dangle': 0,  // Dangling commas are wonderful. It's JSON that's stupid.
-    'func-names': 2,
-    'max-len': 0, // TODO(awong): Enforce after a reindent.
-    'no-console': 2,
-    'no-negated-condition': 0,  // Sometimes negated conditions are easier to understand.
-    'no-unused-expressions': 0, // TODO(awong): Breaks chai expect() statements.
-    'no-unused-vars': [2, {'args': 'after-used', 'argsIgnorePattern': '^_', 'vars': 'local'}],
-    'prefer-rest-params': 0,  // TODO: Enable?
-    'quote-props': [2, 'as-needed', { 'keywords': true }],
-    'require-jsdoc': 0,  // TODO: Enable?
-    'space-unary-ops': 2,
-
-    // Disabled rules with rationale.
-    'react/no-multi-comp': 0,      // Leave organization to code reviewer discretion. 
-    'react/prefer-stateless-function': 0, // Leave statelessness to code reviewer discretion.
-
-    'react/jsx-closing-bracket-location': [2, 'after-props'],
-    'react/jsx-curly-spacing': 2,
-    'react/jsx-equals-spacing': 2,
-    'react/jsx-indent': [2, 2],
-    'react/jsx-key': 2,
-    'react/jsx-no-bind': [2, {'ignoreRefs': true}],
-    'react/jsx-pascal-case': 2,
-    'react/sort-prop-types': [2, {'callbacksLast': true, 'requiredFirst': true}],
-    'react/jsx-sort-props': [2, {'callbacksLast': true, 'shorthandFirst': true}],
-    'react/jsx-handler-names': 2,
-    'react/jsx-indent-props': [2, 4],
-    'react/jsx-no-duplicate-props': 2,
-    'react/no-danger': 2,
-    'react/no-deprecated': 2,
-    'react/no-direct-mutation-state': 2,
-    'react/no-string-refs': 2,
-    'react/prop-types': 2,
-    'react/jsx-space-before-closing': [2, 'never']
-  },
+  'rules': {},
   'parserOptions': {
     'ecmaVersion': 6,
     'sourceType': 'module',
