+++ .travis.yml
@@ -7,6 +7,7 @@ before_script:
 - cp config/saml.key.enc.example config/saml.key.enc
 - cp config/secrets.yml.example config/secrets.yml
 - cp config/application.yml.example config/application.yml
+- npm install
 - bin/rake db:setup --trace
 - bin/rake assets:precompile
 bundler_args: "--deployment --jobs=3 --retry=3 --without deploy development doc"
