+++ app/assets/javascripts/app/utils.js
@@ -1,0 +1,6 @@
+import $ from 'jquery';
+
+
+// Removing element from DOM (on click containing appropriate data attribute)
+const dismiss = '[data-dismiss="true"]';
+$(document).on('click', dismiss, (e) => { $(e.target).parent().remove(); });
