+++ app/assets/javascripts/old/account_type.js
