+++ app/assets/javascripts/old/generating_quiz.js.erb
