+++ app/assets/javascripts/old/idp_help.js.erb
