+++ app/assets/javascripts/old/saml_post_binding.js
