+++ app/assets/stylesheets/components/_close.scss
@@ -1,0 +1,19 @@
+.close {
+  color: currentcolor;
+  float: right;
+  font-size: 20px;
+  line-height: 1;
+}
+
+// Additional properties for button version
+// iOS requires the button element instead of an anchor tag.
+// If you want the anchor version, it requires `href="#"`.
+// See https://developer.mozilla.org/en-US/docs/Web/Events/click#Safari_Mobile
+// scss-lint:disable QualifyingElement
+button.close {
+  -webkit-appearance: none;
+  background: transparent;
+  border: 0;
+  cursor: pointer;
+  padding: 0;
+}
