+++ app/assets/stylesheets/components/all.scss
@@ -1,3 +1,4 @@
 @import 'panel';
 @import 'form';
 @import 'alert';
+@import 'close';
