+++ app/views/layouts/_messages.html.slim
@@ -3,5 +3,8 @@
     - flash.each do |name, msg|
       - if msg.is_a?(String)
         - good_msg = name.to_s.match(/notice|success/)
-        = content_tag :div, msg, class: "p2 alert alert-#{good_msg ? 'success' : 'danger'}", id: "flash_#{name}", tabindex: "#{good_msg ? -1 : 0}"
+        div class="p2 alert alert-#{good_msg ? 'success' : 'danger'}"
+          button type="button" class="close" data-dismiss="true" aria-label="Close"
+            '&times;
+          = msg
     #session_warning_container
