+++ bin/setup
@@ -25,10 +25,9 @@ Dir.chdir APP_ROOT do
 
   puts "\n== Installing dependencies =="
   run "gem install bundler --conservative"
-
   run 'gem install foreman && gem update foreman'
-
   run "bundle check || bundle install"
+  run "npm install"
 
   puts "\n== Preparing database =="
   run "bin/rake db:setup"
