+++ package.json
@@ -2,26 +2,18 @@
   "name": "upaya",
   "version": "0.0.1",
   "private": true,
-  "scripts": {
-    "webpack": "webpack --progress --colors",
-    "dev": "webpack-dev-server --progress --colors",
-    "prod": "webpack -p --progress --colors"
+  "scripts": {},
+  "engines": {
+    "node": "~4.4.x",
+    "npm": "~3.8.x"
   },
   "dependencies": {
-    "react": "^15.0.1",
-    "react-dom": "^15.0.1"
+    "jquery": "^2.2.3"
   },
   "devDependencies": {
-    "babel-core": "^6.7.6",
-    "babel-loader": "^6.2.4",
     "babel-preset-es2015": "^6.6.0",
-    "babel-preset-react": "^6.5.0",
-    "css-loader": "^0.23.1",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "node-sass": "^3.4.2",
-    "sass-loader": "^3.2.0",
-    "style-loader": "^0.13.1",
-    "webpack": "^1.12.15",
-    "webpack-dev-server": "^1.14.1"
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.0",
+    "browserify-incremental": "^3.1.1"
   }
 }
