--- webpack.config.js
@@ -1,29 +1,0 @@
-var webpack = require('webpack');
-
-module.exports = {
-  entry: './public/js/entry.js',
-
-  output: {
-    filename: 'bundle.js',
-    path: __dirname + '/public/build'
-  },
-
-  devtool: '#cheap-module-eval-source-map',
-
-  module: {
-    loaders: [
-      {
-        test: /\.jsx?$/,
-        exclude: /node_modules/,
-        loader: 'babel-loader',
-        query: {
-          presets: ['es2015', 'react'],
-        }
-      },
-    ]
-  },
-
-  resolve: {
-    extensions: ['', '.js', '.jsx']
-  }
-};
